requirements.txt
requests==2.25.1
pymysql==1.0.2
pyyaml==5.4.1
apscheduler==3.7.0

